updates for flow with local dev, testnet dev, or actual modification of meowchain. Issues using ZDC for local deployment after undoing forking on hardhat, use older test helper for now
